+++ src/routes/AppRoutes.tsx
@@ -13,10 +13,12 @@ import UserDashboard from "@src/userPages/UserDashboard";
 import ServicePage from "@src/pages/PointConfig/Service";
 import MemberManagementPage from "@src/pages/MemberManagement";
 import ProfileDetail from "@src/pages/MemberManagement/ProfileDetail";
+import BonusPage from "@src/pages/PointConfig/Bonus";
 import ForgotPasswordPage from "@src/pages/Auth/forgot-password";
 import ProfilePage from "@src/components/templates/Profile";
 import VoucherPage from "@src/pages/Voucher";
 import VoucherAdd from "@src/pages/Voucher/add";
+
 const routes = [
   {
     path: NAVIGATION_ROUTES.DASHBOARD,
@@ -63,6 +65,10 @@ const routes = [
   {
     path: NAVIGATION_ROUTES.PROFILE_DETAIL,
     element: <PrivateRoute Component={ProfileDetail} />,
+  },
+  {
+    path: NAVIGATION_ROUTES.BONUS,
+    element: <PrivateRoute Component={BonusPage} />,
   },
   {
     path: NAVIGATION_ROUTES.FORGOT_PASSWORD,