+++ src/routes/routes.constant.ts
@@ -16,6 +16,7 @@ export const NAVIGATION_ROUTES = {
   SUCCESS: "/success",
   PROFILE_DETAIL: "/profile-detail",
   SERVICE: "/service",
+  BONUS: "/bonus",
   USER_PROFILE: "/profile-edit",
   FORGOT_PASSWORD: "/forgot-password",
   VOUCHER_ADD: "/voucher/add",