+++ CHANGELOG.md
@@ -1,42 +1,57 @@
+<<<<<<< HEAD
+<<<<<<< HEAD
+## 0.0.14 (2023-09-01)
+=======
 ## 0.0.14 (2023-09-07)
+>>>>>>> 6ac5cd9de553439ce0aa32a51895e8809aa6eae6
+=======
+## 0.0.14 (2023-09-07)
+>>>>>>> b544a24e507d413c4540f8da195a6c98243cdd6b
 
 
 ### Bug Fixes
 
 * add page params in property list ([b9be0e0](https://gitlab.com/paras205/boilerplate-react/commit/b9be0e0e3a2c95384720b30b69fead989e878a9a))
 * build issue ([89b7bab](https://gitlab.com/paras205/boilerplate-react/commit/89b7bab7ed5f7fe63870961d67610dc0bfacf3c6))
+<<<<<<< HEAD
 * change list type ([50f0470](https://gitlab.com/paras205/boilerplate-react/commit/50f0470ed7db2a76014e86cf63faf62b18a7538c))
 * colour, design and api response changes ([72990a9](https://gitlab.com/paras205/boilerplate-react/commit/72990a9e3a85297001f6462d54cfa1c5722eb6d3))
 * hotfix: route manage ([9c0394c](https://gitlab.com/paras205/boilerplate-react/commit/9c0394ce422983cccca2b7e439c6f699faf9c135))
 * hotfix: routes ([cdedd07](https://gitlab.com/paras205/boilerplate-react/commit/cdedd07ffe11ad9d5c45c41eedb6d4ac672f82cb))
 * merge ([6445ac1](https://gitlab.com/paras205/boilerplate-react/commit/6445ac124668b2168000f58c6cc81bd9c253ee67))
+=======
+* hotfix: route manage ([9c0394c](https://gitlab.com/paras205/boilerplate-react/commit/9c0394ce422983cccca2b7e439c6f699faf9c135))
+>>>>>>> b544a24e507d413c4540f8da195a6c98243cdd6b
 * refactor code ([7e26a3b](https://gitlab.com/paras205/boilerplate-react/commit/7e26a3bc30eb42c9bb385f334b448c7aaccd916e))
-* remove comment ([af0713c](https://gitlab.com/paras205/boilerplate-react/commit/af0713cfc99343cee95b131be67ba43855644b6d))
 
 
 ### Features
 
 * add additional folder ([6a3e80e](https://gitlab.com/paras205/boilerplate-react/commit/6a3e80e0283bb533435461c064b30935cd953ddf))
 * add and member list completed ([934b4ea](https://gitlab.com/paras205/boilerplate-react/commit/934b4eab6da534d90b24e3b7fe2844ec5de668b4))
+* add date component ([691ecef](https://gitlab.com/paras205/boilerplate-react/commit/691ecefc3ded229b4e421e5e2722d32d98de1c77))
 * added password component, refactored input component ([5676063](https://gitlab.com/paras205/boilerplate-react/commit/5676063ecf247dd38bd7be2ca29c19d3b9c6d76a))
 * added set password template ([75f9471](https://gitlab.com/paras205/boilerplate-react/commit/75f9471f735d8a6cf734924dcc21e0f56326750a))
 * adjusted column width, added edit and add modal in member tier ([827e90e](https://gitlab.com/paras205/boilerplate-react/commit/827e90ebb34d3c5ee73046ad70f05feac8f87be0))
 * admin loagin integartion ([6a58586](https://gitlab.com/paras205/boilerplate-react/commit/6a5858668898a187d75d9633e45aef0ec1867ac1))
-* api integration for add member tier ([591e472](https://gitlab.com/paras205/boilerplate-react/commit/591e4724e8b1fbd33f620dc73ed81df0beed9334))
-* changes on page ([dfb579f](https://gitlab.com/paras205/boilerplate-react/commit/dfb579fe11b8d4d35f8e548ba5798cf50cf145a7))
-* created profile  design changes ([dd38dc0](https://gitlab.com/paras205/boilerplate-react/commit/dd38dc05e611f2eb185ea980105e9d7db10c7fd4))
+* bonus add and list with api ([3496745](https://gitlab.com/paras205/boilerplate-react/commit/34967456fd59e4ccd7fd64c68325373ba949e637))
+* bonus config design and api implement ([dcf5bba](https://gitlab.com/paras205/boilerplate-react/commit/dcf5bba4cb8a4678f5acc5bf6c189cc89b90c925))
 * custom input ([d323c8f](https://gitlab.com/paras205/boilerplate-react/commit/d323c8f84f00f51287e6c564fa41478402ce5ae9))
-* dashboard rewardPoint ([deff013](https://gitlab.com/paras205/boilerplate-react/commit/deff013adbac58d59d98ac6a8241634e040100a7))
-* disabled select if already selected earlier ([17ec76c](https://gitlab.com/paras205/boilerplate-react/commit/17ec76ce3e20f68927377bf1318cee4017490877))
-* forgot password page design and api implementation ([83fdd11](https://gitlab.com/paras205/boilerplate-react/commit/83fdd1147575e825dfdff059120828e07843d96e))
 * header navbar ([db616a8](https://gitlab.com/paras205/boilerplate-react/commit/db616a8aa8ca6b88e5900b684437b8ebf9b3f345))
 * login component ([f753932](https://gitlab.com/paras205/boilerplate-react/commit/f753932002845eaab1196b776d992ee5ecac0547))
 * login component ([967e159](https://gitlab.com/paras205/boilerplate-react/commit/967e159a4fe973d5e6302f108f2bd08525fcc615))
 * login component ([9da8c3e](https://gitlab.com/paras205/boilerplate-react/commit/9da8c3ed3142a97a439caf5080ed56106b689458))
+<<<<<<< HEAD
+<<<<<<< HEAD
+=======
 * logot session clear from font side ([29ac586](https://gitlab.com/paras205/boilerplate-react/commit/29ac5861e27e9b8d2e15e114cc47f0b540ea00ee))
 * password validation ([a14954c](https://gitlab.com/paras205/boilerplate-react/commit/a14954c5935baa2dcd783af303ed3b8ea450aab0))
 * refetched point for manual and added phone validation ([1d51211](https://gitlab.com/paras205/boilerplate-react/commit/1d51211fd510cf093aab1d4739037ff4e30ab572))
+>>>>>>> 6ac5cd9de553439ce0aa32a51895e8809aa6eae6
+=======
+>>>>>>> b544a24e507d413c4540f8da195a6c98243cdd6b
 * resolved modal issue, integrated api for update property ([56fc7b3](https://gitlab.com/paras205/boilerplate-react/commit/56fc7b3dce0062608340685fa86dcee7ab0d45da))
+* started bonus integration ([695c6f0](https://gitlab.com/paras205/boilerplate-react/commit/695c6f003c7cd142db5075ed408956ede0f487a7))